# library_distributed_server/docs/coordinator/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Coordinator service for distributed transactions in the library management system",
        "title": "Distributed Transaction Coordinator API",
        "termsOfService": "N/A",
        "contact": {
            "name": "API Support",
            "email": "dev@library.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/readers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get readers with pagination - THUTHU sees local site only, QUANLY sees all sites",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Readers"
                ],
                "summary": "Get readers based on user role with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number (0-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for reader name or ID",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Readers retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new reader at the user's site (ThuThu only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Readers"
                ],
                "summary": "Create new reader",
                "parameters": [
                    {
                        "description": "Reader information",
                        "name": "reader",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DocGia"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Reader created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.DocGia"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied - ThuThu only",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Reader already exists",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/readers/{maDG}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get reader information by ID - access based on user role and reader location",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Readers"
                ],
                "summary": "Get reader by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reader ID",
                        "name": "maDG",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reader found",
                        "schema": {
                            "$ref": "#/definitions/models.DocGia"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied - reader not at user's site",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Reader not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update reader information (ThuThu only at their site)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Readers"
                ],
                "summary": "Update reader",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reader ID",
                        "name": "maDG",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated reader information",
                        "name": "reader",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DocGia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reader updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.DocGia"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied - ThuThu only or wrong site",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Reader not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a reader (ThuThu only at their site)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Readers"
                ],
                "summary": "Delete reader",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reader ID",
                        "name": "maDG",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reader deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied - ThuThu only or wrong site",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Reader not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Cannot delete reader with active borrows",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DocGia": {
            "description": "Reader information (fragmented by registration branch)",
            "type": "object",
            "required": [
                "hoTen",
                "maCNDangKy",
                "maDG"
            ],
            "properties": {
                "hoTen": {
                    "description": "Reader name",
                    "type": "string",
                    "example": "Nguyễn Văn B"
                },
                "maCNDangKy": {
                    "description": "Registration branch",
                    "type": "string",
                    "example": "Q1"
                },
                "maDG": {
                    "description": "Reader ID",
                    "type": "string",
                    "example": "DG001"
                }
            }
        },
        "models.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "details": {
                    "description": "Error details (optional)"
                },
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "models.ListResponse": {
            "description": "Generic paginated list response matching Flutter BookListModel structure",
            "type": "object",
            "properties": {
                "items": {
                    "description": "List of items (matches Flutter items field)"
                },
                "paging": {
                    "description": "Pagination info (matches Flutter paging field)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PagingInfo"
                        }
                    ]
                }
            }
        },
        "models.PagingInfo": {
            "description": "Pagination information matching Flutter PagingModel structure",
            "type": "object",
            "properties": {
                "page": {
                    "description": "Current page number (0-based, matches Flutter)",
                    "type": "integer",
                    "example": 0
                },
                "size": {
                    "description": "Items per page (matches Flutter)",
                    "type": "integer",
                    "example": 20
                },
                "totalPages": {
                    "description": "Total number of pages (matches Flutter)",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
    }
}